2 out of 3 attempts
When the loop runs for 3 out of 5 attempts the total duration of the test exceeds 40 minutes. This causes an inconvenience to the experimenting subject. According to the Modified Hughson Westlake procedure we can break the loop for 2 positive response out of 3 attempts. By doing so the entire test procedure can be completed in less than 15 minutes with accurate results.